requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
pandas==2.1.0
pytest==7.4.3
python-dotenv==1.0.0